refactor: Views/Models フォルダーに整理し MVVM 構造を導入する (#128)
- MainWindow/AddProfileWindow/ManageProfilesWindow を Views/ に移動
- Models.cs を Models/ に移動
- namespace を XTimelineViewer.Views / XTimelineViewer.Models に変更
- テストプロジェクトのファイルリンクパスを更新
