+++ App.xaml.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Text.Json;
+using XTimelineViewer.Views;
 
 namespace XTimelineViewer
 {
