+++ Services/SettingsService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Text.Json;
+using XTimelineViewer.Models;
 
 namespace XTimelineViewer.Services
 {
