+++ Views/AddProfileWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="XTimelineViewer.AddProfileWindow"
+    x:Class="XTimelineViewer.Views.AddProfileWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     Title="Add Profile">
