+++ Views/MainWindow.Post.cs
@@ -20,7 +20,9 @@ using Windows.Graphics;
 using Windows.Storage;
 using Windows.UI;
 
-namespace XTimelineViewer
+using XTimelineViewer.Models;
+
+namespace XTimelineViewer.Views
 {
     public sealed partial class MainWindow : Window
     {
