+++ Views/MainWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="XTimelineViewer.MainWindow"
+    x:Class="XTimelineViewer.Views.MainWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     Title="XTimelineViewer">
