+++ Views/ManageProfilesWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="XTimelineViewer.ManageProfilesWindow"
+    x:Class="XTimelineViewer.Views.ManageProfilesWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     Title="Manage Profiles">
