+++ XTimelineViewer.Tests/Services/SettingsServiceTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using XTimelineViewer;
+using XTimelineViewer.Models;
 using XTimelineViewer.Services;
 using Xunit;
