+++ XTimelineViewer.Tests/XTimelineViewer.Tests.csproj
@@ -26,7 +26,7 @@
     UI 非依存なファイルだけをリンクして参照する。
   -->
   <ItemGroup>
-    <Compile Include="..\Models.cs"                  Link="Models.cs" />
+    <Compile Include="..\Models\Models.cs"             Link="Models\Models.cs" />
     <Compile Include="..\Services\SettingsService.cs" Link="Services\SettingsService.cs" />
   </ItemGroup>
